refactor: 디렉터리로 변경
