+++ src/routes/index.ts
@@ -1,0 +1,1 @@
+export * from './root'
